requirements.txt
requests==2.18.4
Flask==1.0.2
beautifulsoup4==4.6.3
pyfunctional==1.1.3
lxml==4.2.5